Fix tests so their actions are all in outer classes
