+++ misk/src/test/kotlin/misk/web/WebSocketsTest.kt
@@ -49,26 +49,26 @@ internal class WebSocketsTest {
     )
   }
 
-  @Singleton
-  class EchoWebSocket @Inject constructor() : WebAction {
-    @ConnectWebSocket("/echo")
-    @LogRequestResponse(bodySampling = 1.0, errorBodySampling = 1.0)
-    fun echo(@Suppress("UNUSED_PARAMETER") webSocket: WebSocket): WebSocketListener {
-      return object : WebSocketListener() {
-        override fun onMessage(webSocket: WebSocket, text: String) {
-          webSocket.send("ACK $text")
-        }
-
-        override fun toString() = "EchoListener"
-      }
-    }
-  }
-
   class TestModule : KAbstractModule() {
     override fun configure() {
       install(WebTestingModule())
       install(LogCollectorModule())
       install(WebActionModule.create<EchoWebSocket>())
+    }
+  }
+}
+
+@Singleton
+class EchoWebSocket @Inject constructor() : WebAction {
+  @ConnectWebSocket("/echo")
+  @LogRequestResponse(bodySampling = 1.0, errorBodySampling = 1.0)
+  fun echo(@Suppress("UNUSED_PARAMETER") webSocket: WebSocket): WebSocketListener {
+    return object : WebSocketListener() {
+      override fun onMessage(webSocket: WebSocket, text: String) {
+        webSocket.send("ACK $text")
+      }
+
+      override fun toString() = "EchoListener"
     }
   }
 }
